Folded `From` impls back into mod
